# docs/swagger.yaml
definitions:
  entities.Order:
    properties:
      id:
        type: integer
      product:
        $ref: '#/definitions/entities.Product'
      user:
        $ref: '#/definitions/entities.User'
    type: object
  entities.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      serial_number:
        type: string
    type: object
  entities.User:
    properties:
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/api/orders:
    get:
      consumes:
      - application/json
      description: Get all order information
      responses:
        "200":
          description: array of order with success message
          schema:
            items:
              $ref: '#/definitions/entities.Order'
            type: array
        "201":
          description: null order entities with error message
          schema:
            $ref: '#/definitions/entities.Order'
      summary: Get all order information
      tags:
      - order
    post:
      consumes:
      - application/json
      description: Create for order
      parameters:
      - description: productId
        in: body
        name: product_id
        required: true
        schema:
          type: string
      - description: userId
        in: body
        name: user_id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: order data with success message
          schema:
            $ref: '#/definitions/entities.Order'
        "500":
          description: order null data with failure message
          schema:
            $ref: '#/definitions/entities.Order'
      summary: Insert order info into the database
      tags:
      - order
  /v1/api/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get for order information
      parameters:
      - description: order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: order entities with success message
          schema:
            $ref: '#/definitions/entities.Order'
        "201":
          description: order entities with null value and error message
          schema:
            $ref: '#/definitions/entities.Order'
      summary: Get for order information
      tags:
      - order
...
swagger: "2.0"